chore: cli improvements
